feat: add HeroSection.tsx
